Refactor (APIMF-3491): extracted link generation and id handling pure functions to object
