+++ amf-aml/shared/src/main/scala/amf/aml/internal/parse/instances/DialectInstanceParser.scala
@@ -27,7 +27,12 @@ import amf.aml.internal.parse.common.{AnnotationsParser, DeclarationKey, Declara
 import amf.aml.internal.parse.instances.ClosedInstanceNode.{checkClosedNode, checkRootNode}
 import amf.aml.internal.parse.instances.DialectInstanceParser.{encodedElementDefaultId, findDeclarationsMap, typesFrom}
 import amf.aml.internal.parse.instances.finder.{IncludeFirstUnionElementFinder, JSONPointerUnionFinder}
-import amf.aml.internal.parse.instances.parser.{LiteralCollectionParser, LiteralValueParser, LiteralValueSetter}
+import amf.aml.internal.parse.instances.parser.{
+  ExternalLinkGenerator,
+  LiteralCollectionParser,
+  LiteralValueParser,
+  LiteralValueSetter
+}
 import amf.aml.internal.validate.DialectValidations.{
   DialectAmbiguousRangeSpecification,
   DialectError,
@@ -322,77 +327,7 @@ class DialectInstanceParser(val root: Root)(implicit override val ctx: DialectIn
   }
 
   private def generateExternalLink(id: String, node: YNode, mapping: NodeMapping): Option[DialectDomainElement] = {
-    lazy val instanceTypes = typesFrom(mapping)
-    node.tagType match {
-      case YType.Str => // plain link -> we generate an anonymous node and set the id to the ref and correct type information
-        val elem = DialectDomainElement()
-          .withDefinedBy(mapping)
-          .withId(node)
-          .withIsExternalLink(true)
-          .withInstanceTypes(instanceTypes)
-        Some(elem)
-
-      case YType.Map
-          if node
-            .as[YMap]
-            .key("$id")
-            .isDefined => // simple link in a reference map
-        val refMap = node.as[YMap]
-
-        val id      = explicitNodeId(None, refMap, Nil, "", mapping)
-        val finalId = overrideBase(id, refMap)
-
-        val elem = DialectDomainElement().withDefinedBy(mapping).withId(finalId).withIsExternalLink(true)
-        elem.withInstanceTypes(instanceTypes)
-        elem.annotations += CustomId()
-        refMap.key("$base") match {
-          case Some(baseEntry) =>
-            elem.annotations += CustomBase(baseEntry.value.toString)
-          case _ => // Nothing
-        }
-        Some(elem)
-
-      case YType.Map if mapping.idTemplate.nonEmpty => // complex reference with mandatory idTemplate
-        val refMap = node.as[YMap]
-
-        val element = DialectDomainElement(Annotations(refMap))
-        val id      = idTemplate(element, refMap, Nil, mapping)
-        val finalId = overrideBase(id, refMap)
-
-        // Now we actually parse the provided properties for the node
-        val linkReference: DialectDomainElement =
-          element
-            .withDefinedBy(mapping)
-            .withId(finalId)
-            .withInstanceTypes(instanceTypes)
-            .withIsExternalLink(true) // this is a linkReference
-
-        refMap.key("$base") match {
-          case Some(baseEntry) =>
-            linkReference.annotations += CustomBase(baseEntry.value.toString)
-          case _ => // Nothing
-        }
-
-        // TODO why do we parse properties?
-        mapping.propertiesMapping().foreach { propertyMapping =>
-          val propertyName = propertyMapping.name().value()
-          refMap.key(propertyName) match {
-            case Some(entry) =>
-              parseProperty(finalId, entry, propertyMapping, linkReference)
-            case None => // ignore
-          }
-        }
-
-        // return the parsed reference
-        Some(linkReference)
-
-      case _ => // error
-        ctx.eh.violation(
-            DialectError,
-            id,
-            "AML links must URI links (strings or maps with $id directive) or ID Template links (maps with idTemplate variables)")
-        None
-    }
+    ExternalLinkGenerator.generate(id, node, mapping, root, parseProperty)
   }
 
   private def parseExternalLinkProperty(id: String,
